Prepare backend for frontend integration
 FRONTEND INTEGRATION PREPARATION:

 CORS Configuration:
- Added comprehensive CORS config for all dev ports
- Configurable allowed origins via environment
- Support for credentials and all HTTP methods
- Optimized preflight caching

 API Response Standardization:
- Created ApiResponse<T> wrapper for consistent responses
- Added PaginatedResponse<T> for list endpoints
- Enhanced GlobalExceptionHandler with structured errors
- Added error codes and paths for better frontend handling

 Environment & Documentation:
- Created .env.example template for all environments
- Added comprehensive frontend setup guide
- Enhanced OpenAPI/Swagger configuration
- Added development and production server configs

 FRONTEND-READY FEATURES:
 Standardized error responses with codes
 CORS configured for React/Next.js/Vue
 JWT bearer authentication documented
 Pagination support for data tables
 Environment variable templates
 Complete TypeScript interfaces provided
 API client examples included

 BACKEND IS 100% FRONTEND-READY!

+++ itss-auth-module/src/main/resources/db/migration/V4__create_refresh_tokens_table.sql
@@ -1,7 +1,7 @@
 -- Create refresh_tokens table for managing user sessions
 CREATE TABLE refresh_tokens (
-    id SERIAL PRIMARY KEY,
-    user_id INTEGER NOT NULL REFERENCES users(id) ON DELETE CASCADE,
+    id BIGSERIAL PRIMARY KEY,
+    user_id BIGINT NOT NULL REFERENCES users(id) ON DELETE CASCADE,
     token VARCHAR(500) NOT NULL UNIQUE,
     expires_at TIMESTAMP NOT NULL,
     created_at TIMESTAMP NOT NULL,
